add comments to debounce hook
